backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.26.0
chromadb==0.4.22
langchain==0.1.4
langchain-community==0.0.16
sentence-transformers==2.3.1
gitpython==3.1.41
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
